Fix combat system misbehaving if player is not the first participant.
The combat round will now start immediately when combat begins, without
letting the player select an action (if they participate). Instead,
when the player is reached in the turn order, the round will pause and
UI will be shown so the player can select an action, at which point the
round will resume where it left off.