Implement new string mapping system.
String mappings allow creatures to re-map certain combat-related
strings onto others, enabling different text to be used for a
particular opponent. This system was present in V1 combat but has
now been revamped to work with V2 as well.

Despite the module file version bump, I have not added any upgrade
or backwards-compatibility code for string mappings because they
did not work before and the editor UI for creating them was (and still
is) missing, so it should not be possible for users to have added
them in the first place.