Preventing patch scenes from being loaded (#20)
SceneScripted will now throw an exception if the provided scene file
name cannot be resolved or instantiated for some reason.

It is up to the caller to handle this error in a manner appropriate for
the current context. ExportedSetScene, for example, throws a Lua error.

The game setup logic in GameController.BeginNewSession was left
unprotected because (as far as I'm aware) an invalid scene there
requires save file modification, which is not a supported use case.