Improve scene transition logic and fix memory corruption bug
ExportedSetScene now yields the current coroutine, instead of throwing
an error. That has the same effect in that it returns control to the
SceneScripted but is much faster than preparing and propagating a full-
blown exception for a simple script termination.

Also, there was a bug where, if a particular scene entry point does not
exist, LaunchCoroutine would pop the coroutine, but forget to reset the
m_Coroutine pointer. This could cause Turn() to resume the coroutine,
which could have been GC'd at that point, causing heap corruption.