Fix heap corruption when calling lua_typename.
This was caused by the return value marshalling attribute. The pattern
of UnmanagedType.LPStr is that the CLR assumes ownership of the native
memory block, and subsequently tries to release it. But this isn't
valid for lua_typename since the string is part of ROM, so freeing it
will cause memory corruption.

Instead, there's now a custom marshaller that just takes the pointer
and copies the data into a newly allocated managed string.