Set up framework for script asset specializations.
AssetScript now is only a container for one of 3 ScriptData subtypes:
- External matches the old behavior of a Lua code snippet stored via
  the content attachment interface (i.e. a loose .lua file on disk).
- Inline is a Lua code snippet stored in-file, instead of in a separate
  .lua file on disk like Raw does. Used for scene node scripts, since
  having a loose file for every node is extremely excessive.
- Visual is currently a placeholder, but will contain a graph tree of
  nodes that express an executable script.

Backwards compatibility with furball format V15 is maintained; all
existing scripts will be converted to the Raw or Inline subtypes. This
commit is experimental, and the V16 furball format will change as the
visual scripting feature branch develops.

This change breaks lots of things in the Editor, including all script
editors. These still need to be rewritten to deal with script asset
specializations.