Add a UI for choosing script subtype for new scripts.
When a script is unset, such as with a new item's UseScript, the user
should be able to pick whether they want an External or Visual script.
ScriptEditorHost now allows this by instantiating a NullScriptHost if
there is no ScriptData, which then presents the user with the choice.

The intended content flow for scripts is now such that ScriptData may
be null, but AssetScript may not. For standalone scripts (such as the
item UseScript), the nested AssetScript must be valid, but its child
ScriptData may be null, which indicates an empty script and should be
gracefully handled accordingly.