Refactor AssetSerializer so it discovers all IFurballSerializables in the running assembly dynamically, instead of requiring the static table of types.
This will be much more flexible going forward, since I expect there
will be many, many different subclasses of ScriptNode, and the manual
maintenance on AssetSerializer would become prohibitively tedious.

All found IFurballSerializables will have a constructor lambda
generated for them on the fly, which can be used from then on to
instantiate that type at will. The one annoyance here is that the type
is now serialized as a string even in the binary format, since I can't
piggyback on enums anymore to do compact int serialization, but it's
a small price to pay, all things considered.