Expose the asset registry to ScriptNode code generators.
Some script nodes will have links to other assets, for example a node
for checking whether the player has a certain item, will have to link
to the AssetItem to search for.

ScriptNode.EmitLua now receives an IContentStore interface, implemented
by the ContentStore in the Game project, which allows retrieval of an
asset by its GUID.

The Editor also only needs a GUID to render an AssetPickerControl, so
the ScriptNode will not need to store any more superfluous information.