Support replacing asset instances in AssetWindows.
This finally fixes the broken asset discarding for scenes and string
tables. Before, when making changes to a scene asset and then trying
to discard them by closing the asset window without saving, the
changes would still be committed to disk.

Instead, the scene and string table editors can now duplicate the
source asset when editing begins, and commit a snapshot to the module
whenever the user wishes to save, similar to other assets.