Fix bug where opening >1 tabs for the same asset was possible.
This could lead to all sorts of data loss or data corruption problems,
where the user accidentally has multiple copies of the same asset open
and only one of them is actually committed to disk.

This was a regression introduced by 3fc0f677, where an asset would not
be considered 'open' if its AssetWindow had replaced the instance.
Instead, we now key the open window table with a hash that uses the
asset GUID, so we no longer rely on the actual asset instance pointer.