Also use compressed ints for arrays and byte arrays.
Minor behavioral change with the byte arrays: it is now illegal to write
an empty (zero-length) byte array. (This is to avoid zero-length arrays
becoming `null`s.) Currently the only case this could theoretically
happen is if an item's icon is an empty file masquerading as a PNG.
External scripts write `null` contents when empty or only whitespace, so
they're unaffected.