Allow objects in furball stream to be marked as mandatory
Previously, all objects were optional as far as the parser is concerned.
This increases complexity on the consumer side because it needs to
insert null checks and type checks (which were usually omitted). These
can now be centralized in the reader.

As an extra minor optimization, the binary format can now omit the
'presence' byte for required objects.